llm_caching_service/requirements.txt
fastapi
uvicorn[standard]
gunicorn
redis
pydantic
google-genai
tenacity
PyYAML
structlog
motor
orjson
uvloop
openai
